plotting future trajectory, refactoring and redability changes
