+++ Assets/Scenes/SpaceScene.unity
@@ -227,6 +227,7 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 512236117}
+  - {fileID: 534708853}
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &33588369
@@ -2141,8 +2142,6 @@ MonoBehaviour:
   totalDistanceTravelled: {fileID: 0}
   distanceFromEarth: {fileID: 0}
   distanceFromMoon: {fileID: 0}
-  pastTrajectory: {fileID: 1187994444}
-  futureTrajectory: {fileID: 943471396}
   uiFadeSpeed: 0.4
   inputInactivityTime: 5
   minimumUIVisiblity: 0.2
@@ -2258,6 +2257,52 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 1e8b78ac948f05a46a6d8339a503172b, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!1 &534708852
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 534708853}
+  - component: {fileID: 534708854}
+  m_Layer: 0
+  m_Name: Satellite Manager
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &534708853
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 534708852}
+  serializedVersion: 2
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 17614485}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &534708854
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 534708852}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 680ca114f9c2c8546a1c0d7937b19eba, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  pastTrajectory: {fileID: 1187994444}
+  futureTrajectory: {fileID: 943471396}
 --- !u!1001 &555256269
 PrefabInstance:
   m_ObjectHideFlags: 0
