+++ Assets/Scripts/SatelliteManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 680ca114f9c2c8546a1c0d7937b19eba
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
